use relay hash to find entities

+++ packages/indexer-database/src/entities/HistoricPrice.ts
@@ -16,18 +16,16 @@ export class HistoricPrice {
   @PrimaryGeneratedColumn()
   id: number;
 
-  // bear in mind we are using coingecko symbols directly here, for all intents and purposes this is coingecko historic market price
   @Column()
   baseCurrency: string;
 
   @Column({ default: "usd" })
   quoteCurrency: string;
 
-  // yyyy-LL-dd
   @Column({ type: "date" })
   date: Date;
 
-  @Column({ type: "float" })
+  @Column({ type: "decimal" })
   price: string;
 
   @CreateDateColumn()
